yahoo thingy

+++ src/libraries/requests_util.py
@@ -245,7 +245,7 @@ def get_binance_chart_data(token_name, t_from, t_to):
 
 def get_stock_data(ticker, resolution: int, t_from, t_to):
     resolution_str = str(resolution) + 'm'
-    period = (round(t_to - t_from) / (3600 * 24)) + 1
+    period = int((round(t_to - t_from) / (3600 * 24)) + 1)
     period_str = str(period) + 'd'
     pprint.pprint(period_str)
     pprint.pprint(resolution_str)
